backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
pandas==2.2.1
requests==2.31.0
lxml==5.1.0
python-dateutil==2.9.0
openpyxl==3.1.2
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0
gunicorn==23.0.0
